Resturcture hierarchy
