+++ database/workers/hour_db.bat
@@ -1,0 +1,3 @@
+timeout 15
+python C:\Users\User\Desktop\sunwooang\Hotsan\hourly_main.py
+exit
